LPD-89991 Invoke MCP tools in process to avoid pool exhaustion
The MCP servlet previously made an outbound HTTP call back to its own portal for every tool invocation, which in turn made a second HTTP call to the underlying REST endpoint. Each parallel MCP request therefore held two nested HTTP client connection-pool slots and three Tomcat threads simultaneously, with the outer slot blocked on the inner one. Under enough concurrency this saturated the pool and produced a self-deadlock surfacing as ConnectionPoolTimeoutException for every subsequent call.

The wire-protocol servlet, its supporting classes, and the servlet integration test all move into the REST module so the outer call can be replaced with an in-process invocation of ToolSetUtil. The mcp-server and mcp-server-test bundles are gone; mcp-server-rest-impl and mcp-server-rest-test now hold both halves of the feature.
