--- modules/.releng/apps/mcp/mcp-server/artifact.properties
@@ -1,3 +1,0 @@
-artifact.git.id=1694c20357c66a21ea4e2a70bb9899563a16a590
-artifact.sources.url=https://repository-cdn.liferay.com/nexus/content/repositories/liferay-public-releases/com/liferay/com.liferay.mcp.server/0.0.12/com.liferay.mcp.server-0.0.12-sources.jar
-artifact.url=https://repository-cdn.liferay.com/nexus/content/repositories/liferay-public-releases/com/liferay/com.liferay.mcp.server/0.0.12/com.liferay.mcp.server-0.0.12.jar
