--- modules/.releng/apps/mcp/mcp-server/liferay-releng.changelog
@@ -1,29 +1,0 @@
-#
-# Bundle Version 0.0.2
-#
-48176f256af23e9019370b77d78344f890eab834^..db843ad6c4ba4a562657dc237212465246648004=LPD-66480
-
-#
-# Bundle Version 0.0.4
-#
-db843ad6c4ba4a562657dc237212465246648004^..0fd91974112e608fa9e0b50bd931f7b30224fb95=LPD-66480
-
-#
-# Bundle Version 0.0.6
-#
-0fd91974112e608fa9e0b50bd931f7b30224fb95^..643082b7ba85fc934e1c8d1f8ad402cf610cf652=LPD-77253 LPD-78158
-
-#
-# Bundle Version 0.0.8
-#
-643082b7ba85fc934e1c8d1f8ad402cf610cf652^..e8384f3c8e1a4d5adc26d1adbc60a24a48947353=LPD-73421
-
-#
-# Bundle Version 0.0.10
-#
-e8384f3c8e1a4d5adc26d1adbc60a24a48947353^..11fc62e91e138664b1c5457570339bc30b1ec150=LPD-71114 LPD-83650
-
-#
-# Bundle Version 0.0.12
-#
-11fc62e91e138664b1c5457570339bc30b1ec150^..46ae2b99cc43fe538cd5bf04ab3fab342426afc2=LPD-85549 LPD-86164
