+++ modules/apps/mcp/mcp-server-rest-impl/src/main/java/com/liferay/mcp/server/rest/internal/resource/v1_0/ToolResourceImpl.java
@@ -48,8 +48,7 @@ public class ToolResourceImpl extends BaseToolResourceImpl {
 		}
 
 		return ToolSetUtil.invokeTool(
-			contextHttpServletRequest, object, toolName, toolSetName,
-			contextUriInfo);
+			contextHttpServletRequest, object, toolName, toolSetName);
 	}
 
 }
