+++ modules/apps/mcp/mcp-server-rest-impl/src/main/resources/com/liferay/mcp/server/rest/internal/batch/01-object-definition.batch-engine-data.json
