+++ modules/apps/mcp/mcp-server-rest-impl/src/main/resources/com/liferay/mcp/server/rest/internal/batch/04-list-type-definition.batch-engine-data.json
