+++ modules/apps/mcp/mcp-server-rest-test/bnd.bnd
@@ -1,4 +1,29 @@
 Bundle-Name: Liferay MCP Server REST Test
 Bundle-SymbolicName: com.liferay.mcp.server.rest.test
 Bundle-Version: 1.0.0
--includeresource: com.liferay.mcp.server.rest.client.jar=com.liferay.mcp.server.rest.client-*.jar;lib:=true
+Import-Package:\
+	!com.ethlo.time.*,\
+	\
+	!com.fasterxml.jackson.dataformat.yaml,\
+	\
+	!javax.annotation.*,\
+	\
+	!org.graalvm.*,\
+	\
+	!org.jcodings.*,\
+	\
+	!org.joni.*,\
+	\
+	!org.json.*,\
+	\
+	!org.skyscreamer.*,\
+	\
+	*
+-includeresource:\
+	com.liferay.mcp.server.rest.client.jar=com.liferay.mcp.server.rest.client-*.jar;lib:=true,\
+	reactor-core-[0-9.]*.jar;lib:=true,\
+	@json-[0-9]*.jar;lib:=true,\
+	@json-schema-validator-[0-9.]*.jar;lib:=true,\
+	@jsonassert-[0-9.]*.jar;lib:=true,\
+	@mcp-core-[0-9.]*.jar;lib=true,\
+	@mcp-json-jackson2-[0-9.]*.jar;lib=true
