+++ modules/apps/mcp/mcp-server-rest-test/src/testIntegration/resources/com/liferay/mcp/server/rest/internal/servlet/test/dependencies/post_mcp_server_profiles.json
