--- modules/apps/mcp/mcp-server-test/bnd.bnd
@@ -1,28 +1,0 @@
-Bundle-Name: Liferay MCP Server Test
-Bundle-SymbolicName: com.liferay.mcp.server.test
-Bundle-Version: 1.0.0
-Import-Package:\
-	!com.ethlo.time.*,\
-	\
-	!com.fasterxml.jackson.dataformat.yaml,\
-	\
-	!javax.annotation.*,\
-	\
-	!org.graalvm.*,\
-	\
-	!org.jcodings.*,\
-	\
-	!org.joni.*,\
-	\
-	!org.json.*,\
-	\
-	!org.skyscreamer.*,\
-	\
-	*
--includeresource:\
-	reactor-core-[0-9.]*.jar;lib:=true,\
-	@json-[0-9]*.jar;lib:=true,\
-	@json-schema-validator-[0-9.]*.jar;lib:=true,\
-	@jsonassert-[0-9.]*.jar;lib:=true,\
-	@mcp-core-[0-9.]*.jar;lib=true,\
-	@mcp-json-jackson2-[0-9.]*.jar;lib=true
