--- modules/apps/mcp/mcp-server-test/build.gradle
@@ -1,12 +1,0 @@
-dependencies {
-	testIntegrationImplementation group: "com.fasterxml.jackson.core", name: "jackson-databind", version: "2.18.6"
-	testIntegrationImplementation group: "io.modelcontextprotocol.sdk", name: "mcp-core", version: "1.0.1"
-	testIntegrationImplementation group: "io.modelcontextprotocol.sdk", name: "mcp-json-jackson2", version: "1.0.1"
-	testIntegrationImplementation group: "org.json", name: "json", version: "20231013"
-	testIntegrationImplementation group: "org.skyscreamer", name: "jsonassert", version: "1.5.0"
-	testIntegrationImplementation project(":apps:batch-engine:batch-engine-test-util")
-	testIntegrationImplementation project(":apps:object:object-api")
-	testIntegrationImplementation project(":apps:portal-configuration:portal-configuration-test-util")
-	testIntegrationImplementation project(":apps:portal-vulcan:portal-vulcan-api")
-	testIntegrationImplementation project(":test:arquillian-extension-junit-bridge")
-}
