--- modules/apps/mcp/mcp-server/bnd.bnd
@@ -1,11 +1,0 @@
-Bundle-Name: Liferay MCP Server
-Bundle-SymbolicName: com.liferay.mcp.server
-Bundle-Version: 0.0.13
-Import-Package:\
-	!com.ethlo.*,\
-	\
-	!javax.annotation.*,\
-	\
-	*
--fixupmessages "Classes found in the wrong directory"; restrict:=error; is:=ignore
-Liferay-Client-Extension-Batch: com/liferay/mcp/server/internal/batch
