--- modules/apps/mcp/mcp-server/src/main/java/com/liferay/mcp/server/internal/constants/MCPServerConstants.java
@@ -1,26 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.mcp.server.internal.constants;
-
-/**
- * @author Alejandro Tardín
- */
-public class MCPServerConstants {
-
-	public static final String EXTERNAL_REFERENCE_CODE_MCP_SERVER_DATA_MASK =
-		"L_MCP_SERVER_DATA_MASK";
-
-	public static final String EXTERNAL_REFERENCE_CODE_MCP_SERVER_PROFILE =
-		"L_MCP_SERVER_PROFILE";
-
-	public static final String
-		EXTERNAL_REFERENCE_CODE_MCP_SERVER_PROFILE_DATA_MASK =
-			"L_MCP_SERVER_PROFILE_DATA_MASK";
-
-	public static final String EXTERNAL_REFERENCE_CODE_MCP_SERVER_PROMPT =
-		"L_MCP_SERVER_PROMPT";
-
-}
